Rename o.a.m.cling.invoker.Utils to CliUtils
