+++ impl/maven-cli/src/main/java/org/apache/maven/cling/invoker/BaseParser.java
@@ -54,11 +54,11 @@ import org.apache.maven.properties.internal.EnvironmentUtils;
 import org.apache.maven.properties.internal.SystemProperties;
 
 import static java.util.Objects.requireNonNull;
-import static org.apache.maven.cling.invoker.Utils.getCanonicalPath;
-import static org.apache.maven.cling.invoker.Utils.or;
-import static org.apache.maven.cling.invoker.Utils.prefix;
-import static org.apache.maven.cling.invoker.Utils.stripLeadingAndTrailingQuotes;
-import static org.apache.maven.cling.invoker.Utils.toMap;
+import static org.apache.maven.cling.invoker.CliUtils.getCanonicalPath;
+import static org.apache.maven.cling.invoker.CliUtils.or;
+import static org.apache.maven.cling.invoker.CliUtils.prefix;
+import static org.apache.maven.cling.invoker.CliUtils.stripLeadingAndTrailingQuotes;
+import static org.apache.maven.cling.invoker.CliUtils.toMap;
 
 public abstract class BaseParser implements Parser {
 
@@ -335,7 +335,7 @@ public abstract class BaseParser implements Parser {
 
     @Nullable
     protected Path getRootDirectory(LocalContext context) {
-        return Utils.findRoot(context.topDirectory);
+        return CliUtils.findRoot(context.topDirectory);
     }
 
     protected Map<String, String> populateSystemProperties(LocalContext context) {
