+++ impl/maven-cli/src/main/java/org/apache/maven/cling/invoker/LookupInvoker.java
@@ -94,8 +94,8 @@ import org.slf4j.LoggerFactory;
 import org.slf4j.spi.LocationAwareLogger;
 
 import static java.util.Objects.requireNonNull;
-import static org.apache.maven.cling.invoker.Utils.toMavenExecutionRequestLoggingLevel;
-import static org.apache.maven.cling.invoker.Utils.toProperties;
+import static org.apache.maven.cling.invoker.CliUtils.toMavenExecutionRequestLoggingLevel;
+import static org.apache.maven.cling.invoker.CliUtils.toProperties;
 
 /**
  * Lookup invoker implementation, that boots up DI container.
